Add servers view tests.
