+++ hub/ui/src/views/servers/__tests__/servers.test.jsx
@@ -1,25 +1,40 @@
+jest.mock('../reducer', () => ({
+  defaultState: {
+    hideIdle: false,
+    search: '',
+    sort: 'name',
+  },
+  reducer: jest.fn().mockReturnValue({
+    hideIdle: false,
+    search: '',
+    sort: 'name',
+  }),
+}));
 jest.mock(
   '../../../components/servers-filters',
   () => ({ hideIdle, onFilter, onSearch, onSort, status }) => (
     <>
+      {hideIdle && <div data-testid="hideIdle">hideIdle</div>}
+      <div>{status}</div>
       <button data-testid="btn-filter" onClick={onFilter}>
-        filter
+        Filter
       </button>
-      <button data-testid="btn-onSearch" onClick={onFilter}>
-        filter
+      <button data-testid="btn-onSearch" onClick={() => onSearch('term')}>
+        Search
       </button>
-      <button data-testid="btn-onSort" onClick={onFilter}>
-        filter
+      <button data-testid="btn-onSort" onClick={() => onSort('ready')}>
+        Sort
       </button>
     </>
   ),
 );
 import React from 'react';
-import { render } from '@testing-library/react';
+import { act, fireEvent, render } from '@testing-library/react';
 import '@testing-library/jest-dom/extend-expect';
 import WorkbenchContext from '../../../utils/context';
 
 import Servers from '../';
+import state from '../reducer';
 import { apps, data } from './data';
 
 const Wrapper = props => (
@@ -46,5 +61,53 @@ describe('views/servers', () => {
     );
 
     expect(getByTestId('empty')).toBeInTheDocument();
+  });
+
+  it('should trigger filter action', () => {
+    const dispatch = jest.fn().mockReturnValue(state.defaultState);
+    state.reducer.mockImplementationOnce(dispatch);
+    const { getByTestId } = render(<Wrapper />);
+
+    act(() => {
+      fireEvent.click(getByTestId('btn-filter'));
+    });
+
+    expect(dispatch).toBeCalledWith(state.defaultState, { type: 'toggle' });
+  });
+
+  it('should trigger search', () => {
+    const dispatch = jest.fn().mockReturnValue(state.defaultState);
+    state.reducer.mockImplementationOnce(dispatch);
+    const { getByTestId } = render(<Wrapper />);
+    act(() => {
+      fireEvent.click(getByTestId('btn-onSearch'));
+    });
+
+    expect(dispatch).toHaveBeenLastCalledWith(
+      {
+        hideIdle: false,
+        search: '',
+        sort: 'name',
+      },
+      { type: 'search', payload: 'term' },
+    );
+  });
+
+  it('should trigger search', () => {
+    const dispatch = jest.fn().mockReturnValue(state.defaultState);
+    state.reducer.mockImplementationOnce(dispatch);
+    const { getByTestId } = render(<Wrapper />);
+    act(() => {
+      fireEvent.click(getByTestId('btn-onSort'));
+    });
+
+    expect(dispatch).toHaveBeenLastCalledWith(
+      {
+        hideIdle: false,
+        search: '',
+        sort: 'name',
+      },
+      { type: 'sort', payload: 'ready' },
+    );
   });
 });
