+++ hub/ui/src/views/servers/__tests__/utils.test.js
@@ -1,42 +1,9 @@
-import { defaultState, reducer, processData } from '../utils';
+import { processData } from '../utils';
+import { defaultState } from '../reducer';
 
 import { apps, data } from './data';
 
 describe('views/servers/utils', () => {
-  describe('#reducer', () => {
-    it('should thow without an action type', () => {
-      expect(() => reducer(defaultState, {})).toThrow();
-    });
-
-    it('should toggle', () => {
-      expect(reducer(defaultState, { type: 'toggle' })).toEqual({
-        hideIdle: true,
-        search: '',
-        sort: 'name',
-      });
-    });
-
-    it('should search', () => {
-      expect(
-        reducer(defaultState, { type: 'search', payload: 'test' }),
-      ).toEqual({
-        hideIdle: false,
-        search: 'test',
-        sort: 'name',
-      });
-    });
-
-    it('should sort', () => {
-      expect(reducer(defaultState, { type: 'sort', payload: 'ready' })).toEqual(
-        {
-          hideIdle: false,
-          search: '',
-          sort: 'ready',
-        },
-      );
-    });
-  });
-
   describe('#processData', () => {
     it('should search data', () => {
       const result = processData(
